reclassify richdocument typeerror tests as unit
