+++ test/stdlib/components/docs/test_richdocument.py
@@ -108,23 +108,6 @@ def test_empty_table():
     assert table is None, "table should be empty when supplied string is empty"
 
 
-@pytest.mark.parametrize(
-    "bad_value",
-    ["path/to/file.pdf", "/absolute/path.pdf", 42, None],
-    ids=["str-path", "absolute-str-path", "int", "none"],
-)
-def test_richdocument_init_rejects_non_docling_document(bad_value):
-    with pytest.raises(TypeError, match="DoclingDocument"):
-        RichDocument(bad_value)
-
-
-def test_richdocument_init_rejects_path_object():
-    from pathlib import Path
-
-    with pytest.raises(TypeError, match="from_document_file"):
-        RichDocument(Path("some/file.pdf"))
-
-
 @pytest.mark.skip  # Test requires too much memory for smaller machines.
 @pytest.mark.e2e
 @pytest.mark.huggingface
